Refs #16986 - Header for SyncResultList
